+++ MaceTech.API/Planning/Interfaces/REST/Resources/AssignPlantToDeviceResource.cs
@@ -2,4 +2,4 @@ using System.ComponentModel.DataAnnotations;
 
 namespace MaceTech.API.Planning.Interfaces.REST.Resources;
 
-public record AssignPlantToDeviceResource([Required] int PlantId);
+public record AssignPlantToDeviceResource([Required] long PlantId);
